Finished shaman catalyst

+++ Content/Shaman/OrchidModShamanItem.cs
@@ -8,14 +8,21 @@ using Microsoft.Xna.Framework.Graphics;
 using Terraria.DataStructures;
 using OrchidMod.Common.Attributes;
 using OrchidMod.Common;
+using Terraria.Audio;
 
 namespace OrchidMod.Content.Shaman
 {
 	public enum ShamanCatalystType : int
 	{
 		IDLE = 0,
-		AIM = 1,
-		ROTATE = 2
+		ROTATE = 1
+	}
+
+	public enum ShamanSummonMovement : int
+	{
+		CUSTOM = 0,
+		TOWARDSTARGET = 1,
+		FLOATABOVE = 2
 	}
 
 	[ClassTag(ClassTags.Shaman)]
@@ -24,8 +31,7 @@ namespace OrchidMod.Content.Shaman
 		public ShamanElement Element = ShamanElement.NULL;
 		public Color? catalystEffectColor = null; // TODO: ...
 		public ShamanCatalystType catalystType = ShamanCatalystType.IDLE;
-
-		// ...
+		public ShamanSummonMovement catalystMovement = ShamanSummonMovement.CUSTOM;
 
 		protected override bool CloneNewInstances => true;
 
@@ -81,36 +87,70 @@ namespace OrchidMod.Content.Shaman
 			OrchidShaman shamanPlayer = player.GetModPlayer<OrchidShaman>();
 			if (player.altFunctionUse == 2)
 			{ // Right click
-				if (shamanPlayer.GetShamanicBondValue((ShamanElement)Element) >= 100 && !shamanPlayer.IsShamanicBondReleased((ShamanElement)Element))
+				if (shamanPlayer.GetShamanicBondValue(Element) >= 100 && !shamanPlayer.IsShamanicBondReleased(Element))
 				{
 					shamanPlayer.OnReleaseShamanicBond(this);
-					ShamanElement element = (ShamanElement)Element;
-					switch (element)
+					SoundEngine.PlaySound(SoundID.DD2_GhastlyGlaivePierce, player.Center);
+
+					Vector2 summonPosition = player.Center;
+					Projectile anchor = Main.projectile[shamanPlayer.shamanCatalystIndex];
+					if (anchor.active && anchor.type == ModContent.ProjectileType<CatalystAnchor>())
+					{
+						summonPosition = anchor.Center;
+						anchor.Kill();
+						shamanPlayer.shamanCatalystIndex = -1;
+					}
+
+					var index = Projectile.NewProjectile(null, summonPosition, Vector2.Zero, ModContent.ProjectileType<CatalystSummon>(), (int)player.GetDamage<ShamanDamageClass>().ApplyTo(Item.damage), Item.knockBack, player.whoAmI);
+
+					if (Main.projectile[index].ModProjectile is CatalystSummon catalystSummon)
+					{
+						catalystSummon.SelectedItem = Type;
+						CatalystSummonRelease(Main.projectile[index]);
+						//Main.projectile[index].netUpdate = true; // Unnecessary?
+					}
+					else
+					{
+						Main.NewText("Error creating a new summon catalyst", Color.Red);
+						return false;
+					}
+
+					switch (Element)
 					{
 						case ShamanElement.FIRE:
 							shamanPlayer.ShamanFireBondReleased = true;
 							shamanPlayer.ShamanFireBondPoll = 0;
 							shamanPlayer.ShamanFireBond = shamanPlayer.ShamanBondDuration * 60;
+							shamanPlayer.shamanSummonFireIndex = index;
+							CombatText.NewText(player.Hitbox, ShamanElementUtils.GetColor(ShamanElement.FIRE), "Fire Bond Released");
 							break;
 						case ShamanElement.WATER:
 							shamanPlayer.ShamanWaterBondReleased = true;
 							shamanPlayer.ShamanWaterBondPoll = 0;
 							shamanPlayer.ShamanWaterBond = shamanPlayer.ShamanBondDuration * 60;
+							shamanPlayer.shamanSummonWaterIndex = index;
+							CombatText.NewText(player.Hitbox, ShamanElementUtils.GetColor(ShamanElement.WATER), "Water Bond Released");
 							break;
 						case ShamanElement.AIR:
 							shamanPlayer.ShamanAirBondReleased = true;
 							shamanPlayer.ShamanAirBondPoll = 0;
 							shamanPlayer.ShamanAirBond = shamanPlayer.ShamanBondDuration * 60;
+							shamanPlayer.shamanSummonAirIndex = index;
+							CombatText.NewText(player.Hitbox, ShamanElementUtils.GetColor(ShamanElement.AIR), "Air Bond Released");
 							break;
 						case ShamanElement.EARTH:
 							shamanPlayer.ShamanEarthBondReleased = true;
 							shamanPlayer.ShamanEarthBondPoll = 0;
 							shamanPlayer.ShamanEarthBond = shamanPlayer.ShamanBondDuration * 60;
+							shamanPlayer.shamanSummonEarthIndex = index;
+							CombatText.NewText(player.Hitbox, ShamanElementUtils.GetColor(ShamanElement.EARTH), "Earth Bond Released");
 							break;
 						case ShamanElement.SPIRIT:
 							shamanPlayer.ShamanSpiritBondReleased = true;
 							shamanPlayer.ShamanSpiritBondPoll = 0;
 							shamanPlayer.ShamanSpiritBond = shamanPlayer.ShamanBondDuration * 60;
+							shamanPlayer.shamanSummonSpiritIndex = index;
+							CombatText.NewText(player.Hitbox, ShamanElementUtils.GetColor(ShamanElement.SPIRIT), "Spirit Bond Released");
 							break;
 						default:
 							break;
@@ -177,32 +217,36 @@ namespace OrchidMod.Content.Shaman
 		public sealed override void HoldItem(Player player)
 		{
 			var shaman = player.GetModPlayer<OrchidShaman>();
-			var catalystType = ModContent.ProjectileType<CatalystAnchor>();
 
-			if (player.ownedProjectileCounts[catalystType] == 0)
+			if (!shaman.IsShamanicBondReleased(Element))
 			{
-				var index = Projectile.NewProjectile(null, player.Center, Vector2.Zero, catalystType, 0, 0f, player.whoAmI); // change source to something that's not null ?
-				shaman.shamanCatalystIndex = index;
+				var catalystType = ModContent.ProjectileType<CatalystAnchor>();
 
-				var proj = Main.projectile[index];
-				if (!(proj.ModProjectile is CatalystAnchor catalyst))
+				if (player.ownedProjectileCounts[catalystType] == 0)
 				{
-					proj.Kill();
-					shaman.shamanCatalystIndex = -1;
-				}
-				else catalyst.OnChangeSelectedItem(player);
-			}
-			else
-			{
-				var proj = Main.projectile.First(i => i.active && i.owner == player.whoAmI && i.type == catalystType);
-				if (proj != null && proj.ModProjectile is CatalystAnchor catalyst)
-				{
-					if (catalyst.SelectedItem != player.selectedItem)
+					var index = Projectile.NewProjectile(null, player.Center, Vector2.Zero, catalystType, 0, 0f, player.whoAmI);
+					shaman.shamanCatalystIndex = index;
+
+					var proj = Main.projectile[index];
+					if (!(proj.ModProjectile is CatalystAnchor catalyst))
 					{
-						catalyst.OnChangeSelectedItem(player);
+						proj.Kill();
+						shaman.shamanCatalystIndex = -1;
 					}
+					else catalyst.OnChangeSelectedItem(player);
 				}
-				else shaman.shamanCatalystIndex = -1;
+				else
+				{
+					var proj = Main.projectile.First(i => i.active && i.owner == player.whoAmI && i.type == catalystType);
+					if (proj != null && proj.ModProjectile is CatalystAnchor catalyst)
+					{
+						if (catalyst.SelectedItem != player.selectedItem)
+						{
+							catalyst.OnChangeSelectedItem(player);
+						}
+					}
+					else shaman.shamanCatalystIndex = -1;
+				}
 			}
 
 			SafeHoldItem();
@@ -252,9 +296,10 @@ namespace OrchidMod.Content.Shaman
 		public virtual void PostDrawCatalyst(SpriteBatch spriteBatch, Projectile projectile, Player player, Color lightColor) { }
 		public virtual bool PreAICatalyst(Projectile projectile) { return true; }
 		public virtual bool PreDrawCatalyst(SpriteBatch spriteBatch, Projectile projectile, Player player, ref Color lightColor) { return true; }
-		public virtual void AIReleased(Projectile projectile) { }
-		public virtual void PostAIReleased(Projectile projectile) { }
-		public virtual void OnRelease(Projectile projectile) { }
+		public virtual Vector2 CustomSummonMovementTarget(Projectile projectile) => Main.player[projectile.owner].Center; // Used to create a custom movement target for the summoned catalyst
+		public virtual void CatalystSummonAI(Projectile projectile, int timeSpent) { } //  Used to create custom (attack) ai for the summoned catalyst
+		public virtual void CatalystSummonRelease(Projectile projectile) { } // Called when the summoned catalyst is released (appears)
+		public virtual void CatalystSummonKill(Projectile projectile, int timeSpent) { } // Called when the summoned catalyst is killed
 
 		public virtual string CatalystTexture => "OrchidMod/Content/Shaman/CatalystTextures/" + this.Name + "_Catalyst";
 
@@ -263,6 +308,14 @@ namespace OrchidMod.Content.Shaman
 		public int NewShamanProjectile(Player player, EntitySource_ItemUse source, Vector2 position, Vector2 velocity, int type, int damage, float knockback, float ai0 = 0.0f, float ai1 = 0.0f)
 		{
 			int newProjectileIndex = Projectile.NewProjectile(source, position, velocity, type, damage, knockback, player.whoAmI, ai0, ai1);
+			Projectile newProjectile = Main.projectile[newProjectileIndex];
+			OrchidModProjectile.setShamanBond(newProjectile, (int)Element);
+			return newProjectileIndex;
+		}
+
+		public int NewShamanProjectileFromProjectile(Projectile projectile, Vector2 velocity, int type, int damage, float knockback, float ai0 = 0.0f, float ai1 = 0.0f)
+		{
+			int newProjectileIndex = Projectile.NewProjectile(projectile.GetSource_FromThis(), projectile.Center, velocity, type, damage, knockback, projectile.owner, ai0, ai1);
 			Projectile newProjectile = Main.projectile[newProjectileIndex];
 			OrchidModProjectile.setShamanBond(newProjectile, (int)Element);
 			return newProjectileIndex;
